Drive timer system from game thread
Instead of using a UI timer, the game thread now
drives the timer ticks. For now, it still sends a message
to the UI thread on each tick, so nothing has really changed.
But in the future, Razor will be embedded into the client
instead and then the timer ticks will run inline with the
game loop.