Improve BandageTimer
Make starting and stopping of BandageTimer more robust and not rely on hotkeys: timer will be started as soon as client receives "You begin applying the bandages." message, and stopped when various messages received (added support for "You heal what little damage..." type of messages for the stop event). Also, if displaying of BandageTimer is enabled in settings, explicit "Bandaging started..." and "Bandaging done!" messages will be shown before and after counting.